Add InputOutput wrapper class for testing IO
Formerly, the Doggerel.Exec functions resulted in IO values for printing
output. To avoid having to unit test IO values, generalize execution to
result in a typeclass that allows us to test against a writer monad that
collects the printer output.

It's not yet clear whether this approach will work when we add support
for *input*, but we'll cross that bridge when we come to it.