Statically evaluate all assignments
With this change, assignment commands are evaluated eagerly at the
time that they're executed, rather than at the time that they are
output. The advantage is that we can know dimensionality of
assignments for use in static analysis (to come soon) and we can
emit evaluation errors (such as division by zero) at the location
of the violation.