Store relations in scope with maps
When relations are ready to insert into a scope frame, expand them into
the full maps from a set of units expressions to a pair of the resulting
units and the expression across the key unit set that achieves the value
in those resulting units. This will allow for static analysis of
relation application.