Add support for exponentiation in relations
Relations can now raise values to exponents so long as the radix is a
plain scalar quantity. Extend the expression parser to understand
exponents and encode them in ASTs as unary operators that apply to
vectors.

Although it may be more intuitive to think of exponents as binary
operators, this is done as a concession to relation-isolation simplicity
because, this way, variables can only appear in the mantissa, so
relations can always isolate a value using roots, and do not need to
take logarithms. It should be straightforward to add this support, but,
at this stage, I question whether it's needed.

While the expression parser now understands the exponent operator, it's
only defined in relations' "unit expressions" and an error will be
thrown if an assignment or print expression attempts to use one.