Rework units to record compound dimensionality
Ironing out the semantics doc shows units should be able to record
their compound dimension rather than strictly a base dimension.
Examples of such units might be the unit of Acre: it's
dimensionality is 'area' but that's not a base dimension, it's
length-squared.

This updates the encoding in scope and some matching functionality
to reflect this, but the parser is not updated to accept unit
declarations in non-base dimensions.