Add basic logical operators and boolean dimension
This is a relatively simple start to handling booleans. We define a built-in
dimension of "bool", where we treat vectors of "0 bool" like false and treat
any other value with the same vector dimensionality as false as true. (We can
refer to this dimensionality as "logical dimensionality".)

Define the boolean operators and/or/not and statically check that their operands
are of logical dimensionality.