Add inequality operators
Inequality across vectors is defined when two vectors have the same
dimensionality and the second can be converted to the units of the
first. When they cane, the inequality operator is applied to the
"unit magnitudes" of the vectors.

The idea of a "unit magnitude" of a vector is just the regular
vector magnitude, but in whichever units the vector is measured in.