Lift identifiers used for units and dimensions to static namespace
Closures provide a mechanism for identifiers to go in and out of
scope, but the values they produce in the form of vectors can
be passed through scopes (or they will, when function closures are
supported). Consequently, if some unit X were local, and a vector
in units of X were passed into a context where a different X unit
was local, the meaning of that vector would change, which would be
bad.

Fix this by allowing local dimensions and units, but preventing more
than one from being defined within a single scope frame.