Extend natural unit IDs to their dimension
The dimensionality of a natural unit is now marked with the index of that unit. There's no way to declare an natural dimension, so the only way to achieve a natural dimension in the runtime is to leave a natural unit unassociated with a specific dimension... if that makes sense.

For example, we can declare a natural unit with an association to a dimension like the following.

  unit foo of bar with natural: true;

And consequently, all values in this unit have the same dimensionality.

  > :dims ((1 foo(0)) + (2 foo(1))) + (3 foo(2))
  { bar }

But if we leave the unit unassociated, the its parametric dimension is indexed automatically.

  > unit baz with natural: true;
  > :dims  ((1 baz(0)) + (2 baz(1))) + (3 baz(2))
  { baz(0), baz(1), baz(2) }

And this should be the only way to cause indices to appear on dimensions.