More robust dimspec matching w/ assignment constraints
Upgrades the dimspec matching algorithm to take unbounded ranges
into account (not yet dimspec vars or paren groups) and connect it
up with the assignment constraint mechanism.

Beforehand, assignments could only be constrained to scalars. This
would be like constraining their dimensionality to a dimspec of a
single var.

  let myScalar = expr1 + expr2 with scalar: true;

This will result in an error if the expression results in a multi-
component vector. Now, we can declare a dimspec constraint.

  let myVar = expr1 + expr2 with dims: dim1/dim2^2 + dim3;

Resulting in a similar error if the dimspec does not match the
result.